confirm deletion prompt

+++ pages/single-nachrichten.php
@@ -56,15 +56,12 @@ get_header();
             if ( ( is_user_logged_in() && $current_user->ID == $post->post_author ) ) {
             ?>
                 <a class="button is-style-outline" href="<?php get_permalink(); ?>?action=edit">Nachricht bearbeiten</a>
-                <a class="button is-style-outline" href="<?php get_permalink(); ?>?action=delete">Nachricht löschen</a>
+                <a class="button is-style-outline" onclick="return confirm('Nachricht permanent löschen?')" href="<?php get_permalink(); ?>?action=delete">Nachricht löschen</a>
             <?php
             }
         ?>
 
     </div>
-
-
-
 
 
     <div class="site-content">
